backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
jinja2
python-dotenv
alembic
